Mailer works in one file
