+++ flaskr/__init__.py
@@ -4,7 +4,6 @@ from flask import Flask
 from flask_mail import Mail, Message
 from apscheduler.schedulers.background import BackgroundScheduler
 import os
-
 from flask import Flask, render_template
 
 
@@ -38,28 +37,23 @@ def create_app(test_config=None):
     from . import alerts
     application.register_blueprint(alerts.bp, url_prefix="/alerts")
 
-    # from jobs import send_alert_email
-    # send_alert_email.send_overdue_emails()
-
     logging.basicConfig()
     logging.getLogger('apscheduler').setLevel(logging.DEBUG)
 
     mail = Mail(application)
-
     application.config['MAIL_SERVER'] = 'smtp.gmail.com'
     application.config['MAIL_PORT'] = 465
-    application.config['MAIL_USERNAME'] = 'backStopApp@gmail.com'
-    application.config['MAIL_PASSWORD'] = 'afh8afh8'
+    # import pdb; pdb.set_trace()
+    application.config['MAIL_USERNAME'] = os.getenv('MAIL_USERNAME')
+    application.config['MAIL_PASSWORD'] = os.getenv('MAIL_PASSWORD')
     application.config['MAIL_USE_TLS'] = False
     application.config['MAIL_USE_SSL'] = True
     mail = Mail(application)
-
 
     def send_overdue_emails():
         with application.app_context():
             now = datetime.now()
             database = db.get_db()
-
             alerts = database.execute(
                 'SELECT id, title, schedule, email, message, sent'
                 ' FROM alert'
@@ -67,22 +61,20 @@ def create_app(test_config=None):
             ).fetchall()
 
             for alert in alerts:
-                print(alert['id'])
                 msg = Message(alert['title'], sender='backStopApp@gmail.com',
                             recipients=[alert['email']])
-                
                 msg.body = alert['message']
                 mail.send(msg)
+                database.execute(
+                    f"UPDATE alert set sent = true WHERE id = {alert['id']}"
+                )
                 return "Sent"
-
 
     scheduler = BackgroundScheduler()
     scheduler.add_job(send_overdue_emails, 'cron', minute='*/1')
     scheduler.start()
 
     return application
-
-
