+++ flaskr/db.py
@@ -9,7 +9,7 @@ def get_db():
     if 'db' not in g:
         g.db = sqlite3.connect(
             current_app.config['DATABASE'],
-            detect_types=sqlite3.PARSE_DECLTYPES
+            detect_types=(sqlite3.PARSE_DECLTYPES | sqlite3.PARSE_COLNAMES)
         )
         g.db.row_factory = sqlite3.Row
